bluemix-python-flask-sample-master/requirements.txt
Flask==0.10.1
python-keystoneclient==1.7.1
python-swiftclient==2.6.0
couchdb